Make useAdapter customizable
