+++ package.json
@@ -45,6 +45,9 @@
     "watch:all": "node scripts/watch-all.js",
     "watch:venia": "yarn venia run watch"
   },
+  "dependencies": {
+    "caniuse-lite": "~1.0.30001335"
+  },
   "devDependencies": {
     "@babel/plugin-transform-runtime": "~7.4.4",
     "@graphql-cli/validate": "~2.1.0",
